Create SearchInRange.java
You are given an array of integers and a target integer. Your task is to implement the linear search algorithm to find the target element within a specified range in the array.

If the target is found within the given range, return its index.
If the target is not found within the given range, return -1.
The search should be performed only between the given start and end indices (both inclusive).